Added initial version of modified script.
Now the script will:
Switch to A partition first before flashing, then to B, and finally back to A.
Explicitly call fastboot from the directory I have fastboot installed in
Prepend fastboot command with sudo in order for the fastboot command to work with my setup.
Added prompt after flashing images to reboot into system or stay in fastboot mode.